genai/langchain/chatbot/requirements.txt
langchain
openai
hugginface_hub
python-dotenv
streamlit